fix: auth 헬퍼를 별도 파일로 분리하여 빌드 문제 해결
auth.ts에서 @sori/database를 직접 import하면 클라이언트 라우트에서
import 시 Vite가 전체 database 패키지를 분석하여 SSR 빌드 실패.

- auth-helpers.ts로 권한 검증 헬퍼 분리
- 동적 import 사용으로 번들 분리
- SSR 모듈 수: 2894개 → 109개로 정상화
